requirements.txt
Django==3.2.15
django-filter
django-environ==0.8.1
psycopg2-binary==2.9.3
djangorestframework==3.14.0